+++ cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # cline
 
+## 2.4.2
+
+### Added
+
+- Gemini-3.1 Pro Preview
+
+### Patch Changes
+
+- VSCode uses shared files for global, workspace and secret state.
+
 ## [2.4.1]
 
 ### Fixed